+++ envcoord/mailexpander/expander.py
@@ -15,7 +15,6 @@ import getopt
 import ldap
 import logging
 import os
-import re
 import smtplib
 import sys
 import time
@@ -290,6 +289,10 @@ class Expander(object):
             if self.also_send_to != ['']:
                 retval = self.send_emails(
                     role_email, self.also_send_to, content)
+                if retval != RETURN_CODES['EX_OK']:
+                    log.error("Error %s while sending to %s",
+                              retval, self.also_send_to)
+                    return retval
 
             # Send e-mails
             for emails in email_batches:
@@ -297,7 +300,7 @@ class Expander(object):
                     role_email, emails, content)
                 if retval != RETURN_CODES['EX_OK']:
                     log.error("Error %s while sending to %s",
-                              retval, self.no_owner_send_to)
+                              retval, emails)
                     return retval
             if not self.skip_confirmation_email:
                 try:
@@ -378,8 +381,8 @@ class Expander(object):
         # this allows "inheriting" permissions from above roles
         # see http://taskman.eionet.europa.eu/issues/20422
 
-        if 'permittedperson' not in role_data:
-            role_data['permittedperson'] = []
+        if 'permittedPerson' not in role_data:
+            role_data['permittedPerson'] = []
         if 'permittedSender' not in role_data:
             role_data['permittedSender'] = []
 
@@ -418,13 +421,13 @@ class Expander(object):
 
             for person_dn in role_info.get('permittedPerson', []):
                 try:
-                    email = self.agent._query(person_dn)['mail'][0]
+                    sender_email = self.agent._query(person_dn)['mail'][0]
                 except Exception as e:
                     # Log that we couldn't get the email.
                     log.exception("Invalid DN: %s; %s" % (person_dn, e))
                     continue
                 else:
-                    senders.add(email)
+                    senders.add(sender_email)
 
         role_data['permittedSender'] = list(filter(None, set(senders)))